==> xenium_HYP/code/07_Domains-subdomains_by_knnSmoothing_Banksytypes/05b-Glial_proximityAndDensity_vs_each_VMHARC_clus.Rmd <==
---
title: "05b-Glial_proximityAndDensity_vs_each_VMHARC_clus"
output: html_document
date: "2024-09-24"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.height = 10,fig.width = 7,include = FALSE)
#### sets tab autocompletion for directories to begin from the directory containing the .Rproj session, instead of the script's directory if different
knitr::opts_knit$set(root.dir = here::here())

library(data.table)
library(ggplot2)
library(gridExtra)
library(SpatialExperiment)
library(SpatialFeatureExperiment)
library(sf)

## rstudio GUI tweaks
require(colorout)
ColorOut()
options("styler.addins_style_transformer" = "biocthis::bioc_style()")
##

## enable forked parallel processing with BiocParallel::multicoreParam, future::, etc. seemed to need this a couple times, but otherwise havent so its here as a preventative measure. part of this is adding the line 
# OBJC_DISABLE_INITIALIZE_FORK_SAFETY=YES
# to Renviron.site. see e.g. top response on https://stackoverflow.com/questions/73638290/python-on-mac-is-it-safe-to-set-objc-disable-initialize-fork-safety-yes-globall 
library(parallelly)
options(parallelly.supportsMulticore.disableOn="")
options(parallelly.fork.enable=TRUE)
library(BiocParallel)
options(bphost="localhost")
library(parallel)


## ggplot defaults
theme_set(theme_bw()+theme(axis.text.x = element_text(size = 14), axis.title.x = element_text(size = 16), axis.text.y = element_text(size = 14), axis.title.y = element_text(size =16), plot.title = element_text(size = 20,hjust=0.5), strip.text = element_text(size=18), legend.text = element_text(size=10), legend.title = element_text(size=11,hjust=0.5)))

## last of all, unload the base package datasets, whose data keep getting in the way of autocompletions (e.g., Theoph is priortized over TRUE)
unloadNamespace("datasets")
```

test 1: local proportion of cells that are of each glial supertype or of any glial supertype (get all cells within a couple different radii of a cell from each vmh/arc cluster, proportion of those that have glial label to start)

test 2: are the glial-crossover genes those most highly expressed in glial types surrounding ARC-glia or VMH-glia crossover pops? (for any glial types identified by ^)

### data setup:
```{r}
hypx <- readRDS("processed-data/05_Banksy_M0lam0_res2_multisamp/01-sfe-in_staggered-coords_genetarg-only_NONlog-norm.RDS")

bkcl <- fread("processed-data/05_Banksy_M0lam0_res2_multisamp/01-BanksyClusts_M0lam0_leiden_multisamp.txt")
bkanno <- fread("processed-data/05_Banksy_M0lam0_res2_multisamp/02-M0l0kg6_topClusMarkers_celltypes_annotated.txt")

## merge annotations to raw banksy cluster-cell-assignment output; drop sample specific clusters (DISCARD)
bkanno <- unique(bkanno[,.(clus,bjm_annot)])
bkanno <- bkanno[!(bjm_annot %in% grep(bjm_annot, pattern="DISCARD",value=T))]

setnames(bkcl,2,"clus")
bkcl <- merge(bkcl,bkanno,by="clus")

## subset SPE to cells not labeled as donor spec clust
hypx <- hypx[,bkcl$rn]

## carefully, probably excessively, assign annotated cellc lusters to coldata
bkcl <- DataFrame(bkcl,row.names=bkcl$rn)[colnames(hypx),]
colLabels(hypx) <- bkcl$bjm_annot

## now append the domain definitions
smoove <- fread("processed-data/07_Domains-subdomains_by_knnSmoothing_Banksytypes/03b-ARCVMHdomains_2stepsmooth_VMH-k10-0.2-VMH-k200-0.2_ARC-k50-0.1_ARC-k500-0.5.txt")
smoove <- smoove[rn %in% colnames(hypx)]
smoove <- DataFrame(smoove,row.names=smoove$rn)[colnames(hypx),]
hypx$smoovedom <- smoove$dualVMHARC4

## for computational effish subset to ARC and VMH cells within their respective domains
hyparc <- hypx[,hypx$smoovedom=="ARC"]
hypvmh <- hypx[,hypx$smoovedom=="VMH"]

rm(hypx)
gc(full=T)
```

## find how many cells are within 50, 100, 200, 300, 400, 500 units of a given cell within VMH
```{r}
vmhlocs <- as.data.table(cbind(spatialCoords(hypvmh),as.data.table(colData(hypvmh))[,.(sample_id,key,label)]))

samps <- unique(vmhlocs$sample_id)
vmhdists <- mclapply(samps,mc.cores=8,FUN=function(s){
    vmhlocs.sf <- st_as_sf(vmhlocs[sample_id==s],coords=c("sdimx","sdimy"))
    dists <- c(50,100,200,300,400,500)
    sampres <- lapply(dists,FUN=function(d){
        return(st_is_within_distance(vmhlocs.sf,vmhlocs.sf,dist=d))
    })
    names(sampres) <- c("d50","d100","d200","d300","d400","d500")
    return(sampres)
})
names(vmhdists) <- samps

## extract the pair IDs. the names of the two columns turn out differently for each sublist entry (radius length), so use .SDcols=1 to get the first column (pair index 1) or =2 for pair index 2 in the calls to lapply(.SD) below
vmhdists2 <- mapply(v=vmhdists,s=samps,SIMPLIFY=FALSE,FUN=function(v,s){
    vmhlocs.s <- vmhlocs[sample_id==s]
    
    ret <- lapply(v,function(rad){
        rad2 <- as.data.table(as.data.frame(rad))
        rad2[,cell1:=lapply(.SD,FUN=function(x){vmhlocs.s[x,key]}),.SDcols=1]
        rad2[,cell2:=lapply(.SD,FUN=function(x){vmhlocs.s[x,key]}),.SDcols=2]
        
        # drop self-pairs
        rad2 <- rad2[cell1!=cell2]
        
        # add cell label info now that we have cell ids for each cell in each pair
        z <- merge.data.table(rad2,vmhlocs[sample_id==s,.(key,label)],by.x="cell1",by.y="key")
        setnames(z,"label","label1")
        a <- merge.data.table(z,y=vmhlocs[sample_id==s,.(key,label)],by.x="cell2",by.y="key")
        setnames(a,"label","label2")
        a[,sample_id:=s]
        
        ## number of NEARBY cells of a given type (label2) for a given cell (cell1). also return the label of the given cell (label1) so we can filter these by target cell type later. 
        nperclus <- a[,.(.N,label1),by=c("cell1","label2")]
        return(nperclus)
    })
    return(ret)
})

rm(vmhdists,hypvmh,vmhlocs)
gc(full=T)

## save the intermediate object with tables of cell type pairs and labels
saveRDS(vmhdists2,"processed-data/07_Domains-subdomains_by_knnSmoothing_Banksytypes/05b-VMHdomain_cellpairs_with_labels_100-500u_and_50u_radii.RDS")

## this is a pretty big file, so remove from workspace for now to run thru ARC. then load them back up one at a time for downstream stuff.
rm(vmhdists2)
gc(full=T)
```

## find how many cells are within 50, 100, 200, 300, 400, 500 units of a given cell within ARC

```{r}
arclocs <- as.data.table(cbind(spatialCoords(hyparc),as.data.table(colData(hyparc))[,.(sample_id,key,label)]))


arcdists <- mclapply(samps,mc.cores=8,FUN=function(s){
    arclocs.sf <- st_as_sf(arclocs[sample_id==s],coords=c("sdimx","sdimy"))
    dists <- c(50,100,200,300,400,500)
    sampres <- lapply(dists,FUN=function(d){
        return(st_is_within_distance(arclocs.sf,arclocs.sf,dist=d))
    })
    names(sampres) <- c("d50","d100","d200","d300","d400","d500")
    return(sampres)
})
names(arcdists) <- samps


## extract the pair IDs. the names of the two columns turn out differently for each sublist entry (radius length), so use .SDcols=1 to get the first column (pair index 1) or =2 for pair index 2 in the calls to lapply(.SD) below
arcdists2 <- mapply(v=arcdists,s=samps,SIMPLIFY=FALSE,FUN=function(v,s){
    arclocs.s <- arclocs[sample_id==s]
    
    ret <- lapply(v,function(rad){
        rad2 <- as.data.table(as.data.frame(rad))
        rad2[,cell1:=lapply(.SD,FUN=function(x){arclocs.s[x,key]}),.SDcols=1]
        rad2[,cell2:=lapply(.SD,FUN=function(x){arclocs.s[x,key]}),.SDcols=2]
        
        # drop self-pairs
        rad2 <- rad2[cell1!=cell2]
        
        # add cell label info now that we have cell ids for each cell in each pair
        z <- merge.data.table(rad2,arclocs[sample_id==s,.(key,label)],by.x="cell1",by.y="key")
        setnames(z,"label","label1")
        a <- merge.data.table(z,y=arclocs[sample_id==s,.(key,label)],by.x="cell2",by.y="key")
        setnames(a,"label","label2")
        a[,sample_id:=s]
        
        ## number of NEARBY cells of a given type (label2) for a given cell (cell1). also return the label of the given cell (label1) so we can filter these by target cell type later. 
        nperclus <- a[,.(.N,label1),by=c("cell1","label2")]
        return(nperclus)
    })
    return(ret)
})

rm(arcdists,hyparc,arclocs)
gc(full=T)

## save the intermediate object with tables of cell type pairs and labels
saveRDS(arcdists2,"processed-data/07_Domains-subdomains_by_knnSmoothing_Banksytypes/05b-ARCdomain_cellpairs_with_labels_100-500u_and_50u_radii.RDS")
```


Now, (starting with the arc ones, which we didn't remove from memory yet), subset down to cells in each ARC cluster and determine the proportion of neighbors that are every other cell type. Then load baack in VMH results and do  the same.
```{r}
arccelldists <- lapply(arcdists2,FUN=function(ad){
    lapply(ad,FUN=function(adr){
        adr <- adr[label1 %in% paste0("ARC_",c(1:5))]
        return(adr)
    })
})
rm(arcdists2)
gc(full=T)

## now tabulate how many total neighboring cells there are and the proportion of those w a given label
arccelldists <- mclapply(arccelldists,mc.cores=5,FUN=function(ad){
    lapply(ad,FUN=function(adr){
        adr[,total_cell1_neighbors:=.N,by="cell1"]
        adr[,prop_celltype_inproximity:=.N/total_cell1_neighbors,by=c("cell1","label2")]
        ## we created a redundant-ish row above named N that's just the number of cells of a given type in proximity to a given cell. but the only things we need now are label1, cell1, label2, N, total in proximity, and prop in proximity.
        adr <- adr[,.(label1,cell1,label2,total_cell1_neighbors,N,prop_celltype_inproximity)]
        setnames(adr,"N","N_label2_neighboring_cell1")
        return(unique(adr))
    })
})

# save
saveRDS(arccelldists,"processed-data/07_Domains-subdomains_by_knnSmoothing_Banksytypes/05b-ARCdomain_ARCcells_NandProp_pertype_within_100-500u_and_50u_radii.RDS")

rm(arccelldists)
```


#### now for VMH:
```{r}
vmhdists2 <- readRDS("processed-data/07_Domains-subdomains_by_knnSmoothing_Banksytypes/05b-VMHdomain_cellpairs_with_labels_100-500u_and_50u_radii.RDS")

vmhcelldists <- lapply(vmhdists2,FUN=function(vd){
    lapply(vd,FUN=function(vdr){
        vdr <- vdr[label1 %in% paste0("VMH_",c(1:3,"Lateral_border"))]
        return(vdr)
    })
})

rm(vmhdists2)
gc(full=T)

## now tabulate how many total neighboring cells there are and the proportion of those w a given label
vmhcelldists <- mclapply(vmhcelldists,mc.cores=5,FUN=function(vd){
    lapply(vd,FUN=function(vdr){
        vdr[,total_cell1_neighbors:=.N,by="cell1"]
        vdr[,prop_celltype_inproximity:=.N/total_cell1_neighbors,by=c("cell1","label2")]
        ## we created a redundant-ish row above named N that's just the number of cells of a given type in proximity to a given cell. but the only things we need now are cell1, label2, N, total in proximity, and prop in proximity.
        vdr <- vdr[,.(label1,cell1,label2,total_cell1_neighbors,N,prop_celltype_inproximity)]
        setnames(vdr,"N","N_label2_neighboring_cell1")
        return(unique(vdr))
    })
})


# save
saveRDS(vmhcelldists,"processed-data/07_Domains-subdomains_by_knnSmoothing_Banksytypes/05b-VMHdomain_VMHcells_NandProp_pertype_within_100-500u_and_50u_radii.RDS")
```

reprod info:

```{r}
sessionInfo()
sessioninfo::session_info()
```

R version 4.4.1 RC (2024-06-06 r86719)
Platform: aarch64-apple-darwin20
Running under: macOS Sonoma 14.5

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib 
LAPACK: /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.0

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

time zone: America/Chicago
tzcode source: internal

attached base packages:
[1] parallel  stats4    stats     graphics  grDevices utils     methods   base     

other attached packages:
 [1] edgeR_4.2.0                    dreamlet_1.2.1                
 [3] variancePartition_1.34.0       limma_3.60.3                  
 [5] sf_1.0-16                      BiocParallel_1.38.0           
 [7] parallelly_1.38.0              colorout_1.3-0.2              
 [9] spatialLIBD_1.16.2             SpatialFeatureExperiment_1.6.1
[11] SpatialExperiment_1.14.0       SingleCellExperiment_1.26.0   
[13] SummarizedExperiment_1.34.0    Biobase_2.64.0                
[15] GenomicRanges_1.56.1           GenomeInfoDb_1.40.1           
[17] IRanges_2.38.1                 S4Vectors_0.42.1              
[19] BiocGenerics_0.50.0            MatrixGenerics_1.16.0         
[21] matrixStats_1.3.0              gridExtra_2.3                 
[23] ggplot2_3.5.1                  data.table_1.15.4             
[25] rlang_1.1.4                   

loaded via a namespace (and not attached):
  [1] bitops_1.0-7              EBImage_4.46.0            httr_1.4.7               
  [4] RColorBrewer_1.1-3        doParallel_1.0.17         Rgraphviz_2.48.0         
  [7] numDeriv_2016.8-1.1       backports_1.5.0           tools_4.4.1              
 [10] metafor_4.6-0             utf8_1.2.4                R6_2.5.1                 
 [13] DT_0.33                   HDF5Array_1.32.0          lazyeval_0.2.2           
 [16] rhdf5filters_1.16.0       withr_3.0.0               sp_2.1-4                 
 [19] prettyunits_1.2.0         cli_3.6.3                 sass_0.4.9               
 [22] KEGGgraph_1.64.0          SQUAREM_2021.1            mvtnorm_1.2-5            
 [25] proxy_0.4-27              mixsqp_0.3-54             Rsamtools_2.20.0         
 [28] R.utils_2.12.3            zenith_1.6.0              scater_1.32.0            
 [31] sessioninfo_1.2.2         attempt_0.3.1             invgamma_1.1             
 [34] maps_3.4.2                rstudioapi_0.16.0         RSQLite_2.3.7            
 [37] generics_0.1.3            BiocIO_1.14.0             gtools_3.9.5             
 [40] spdep_1.3-5               dplyr_1.1.4               metadat_1.2-0            
 [43] Matrix_1.7-0              ggbeeswarm_0.7.2          fansi_1.0.6              
 [46] abind_1.4-5               R.methodsS3_1.8.2         terra_1.7-78             
 [49] lifecycle_1.0.4           yaml_2.3.9                mathjaxr_1.6-0           
 [52] gplots_3.1.3.1            rhdf5_2.48.0              SparseArray_1.4.8        
 [55] BiocFileCache_2.12.0      paletteer_1.6.0           grid_4.4.1               
 [58] blob_1.2.4                promises_1.3.0            dqrng_0.4.1              
 [61] ExperimentHub_2.12.0      crayon_1.5.3              lattice_0.22-6           
 [64] msigdbr_7.5.1             beachmat_2.20.0           cowplot_1.1.3            
 [67] annotate_1.82.0           KEGGREST_1.44.1           magick_2.8.3             
 [70] zeallot_0.1.0             pillar_1.9.0              knitr_1.48               
 [73] rjson_0.2.21              boot_1.3-30               corpcor_1.6.10           
 [76] codetools_0.2-20          wk_0.9.2                  glue_1.7.0               
 [79] vctrs_0.6.5               png_0.1-8                 spam_2.10-0              
 [82] Rdpack_2.6                gtable_0.3.5              assertthat_0.2.1         
 [85] rematch2_2.1.2            cachem_1.1.0              xfun_0.45                
 [88] rbibutils_2.2.16          S4Arrays_1.4.1            mime_0.12                
 [91] DropletUtils_1.24.0       Rfast_2.1.0               sfheaders_0.4.4          
 [94] iterators_1.0.14          units_0.8-5               fields_16.2              
 [97] statmod_1.5.0             nlme_3.1-165              pbkrtest_0.5.3           
[100] bit64_4.0.5               progress_1.2.3            EnvStats_2.8.1           
[103] filelock_1.0.3            rprojroot_2.0.4           bslib_0.7.0              
[106] irlba_2.3.5.1             vipor_0.4.7               KernSmooth_2.23-24       
[109] rmeta_3.0                 colorspace_2.1-0          spData_2.3.1             
[112] DBI_1.2.3                 tidyselect_1.2.1          bit_4.0.5                
[115] compiler_4.4.1            curl_5.2.1                graph_1.82.0             
[118] BiocNeighbors_1.22.0      DelayedArray_0.30.1       plotly_4.10.4            
[121] rtracklayer_1.64.0        scales_1.3.0              caTools_1.18.2           
[124] remaCor_0.0.18            classInt_0.4-10           rappdirs_0.3.3           
[127] stringr_1.5.1             tiff_0.1-12               digest_0.6.36            
[130] fftwtools_0.9-11          minqa_1.2.7               rmarkdown_2.27           
[133] benchmarkmeData_1.0.4     aod_1.3.3                 XVector_0.44.0           
[136] RhpcBLASctl_0.23-42       htmltools_0.5.8.1         pkgconfig_2.0.3          
[139] jpeg_0.1-10               lme4_1.1-35.5             sparseMatrixStats_1.16.0 
[142] mashr_0.2.79              dbplyr_2.5.0              fastmap_1.2.0            
[145] htmlwidgets_1.6.4         UCSC.utils_1.0.0          shiny_1.8.1.1            
[148] DelayedMatrixStats_1.26.0 jquerylib_0.1.4           jsonlite_1.8.8           
[151] config_0.3.2              R.oo_1.26.0               BiocSingular_1.20.0      
[154] RCurl_1.98-1.14           magrittr_2.0.3            scuttle_1.14.0           
[157] GenomeInfoDbData_1.2.12   s2_1.1.6                  dotCall64_1.1-1          
[160] Rhdf5lib_1.26.0           munsell_0.5.1             Rcpp_1.0.13              
[163] babelgene_22.9            viridis_0.6.5             RcppZiggurat_0.1.6       
[166] EnrichmentBrowser_2.34.1  stringi_1.8.4             zlibbioc_1.50.0          
[169] MASS_7.3-61               plyr_1.8.9                AnnotationHub_3.12.0     
[172] ggrepel_0.9.5             deldir_2.0-4              Biostrings_2.72.1        
[175] splines_4.4.1             hms_1.1.3                 locfit_1.5-9.10          
[178] reshape2_1.4.4            ScaledMatrix_1.12.0       BiocVersion_3.19.1       
[181] XML_3.99-0.17             evaluate_0.24.0           golem_0.4.1              
[184] RcppParallel_5.1.8        BiocManager_1.30.23       nloptr_2.1.1             
[187] foreach_1.5.2             httpuv_1.6.15             tidyr_1.3.1              
[190] purrr_1.0.2               scattermore_1.2           benchmarkme_1.0.8        
[193] ashr_2.2-63               rsvd_1.0.5                broom_1.0.6              
[196] xtable_1.8-4              restfulr_0.0.15           fANCOVA_0.6-1            
[199] e1071_1.7-14              later_1.3.2               viridisLite_0.4.2        
[202] class_7.3-22              truncnorm_1.0-9           tibble_3.2.1             
[205] lmerTest_3.1-3            memoise_2.0.1             beeswarm_0.4.0           
[208] AnnotationDbi_1.66.0      GenomicAlignments_1.40.0  shinyWidgets_0.8.6       
[211] GSEABase_1.66.0           here_1.0.1               
> sessioninfo::session_info()
─ Session info ───────────────────────────────────────────────────────────────────
 setting  value
 version  R version 4.4.1 RC (2024-06-06 r86719)
 os       macOS Sonoma 14.5
 system   aarch64, darwin20
 ui       RStudio
 language (EN)
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       America/Chicago
 date     2024-09-25
 rstudio  2024.07.0-daily+219 Cranberry Hibiscus (desktop)
 pandoc   3.1.11 @ /Applications/RStudio.app/Contents/Resources/app/quarto/bin/tools/aarch64/ (via rmarkdown)

─ Packages ───────────────────────────────────────────────────────────────────────
 ! package                  * version    date (UTC) lib source
   abind                      1.4-5      2016-07-21 [1] CRAN (R 4.4.0)
   annotate                   1.82.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   AnnotationDbi              1.66.0     2024-05-01 [1] Bioconductor 3.19 (R 4.4.0)
   AnnotationHub              3.12.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   aod                        1.3.3      2023-12-13 [1] CRAN (R 4.4.0)
   ashr                       2.2-63     2023-08-21 [1] CRAN (R 4.4.0)
   assertthat                 0.2.1      2019-03-21 [1] CRAN (R 4.4.0)
   attempt                    0.3.1      2020-05-03 [1] CRAN (R 4.4.0)
   babelgene                  22.9       2022-09-29 [1] CRAN (R 4.4.0)
   backports                  1.5.0      2024-05-23 [1] CRAN (R 4.4.0)
   beachmat                   2.20.0     2024-05-06 [1] Bioconductor 3.19 (R 4.4.0)
   beeswarm                   0.4.0      2021-06-01 [1] CRAN (R 4.4.0)
   benchmarkme                1.0.8      2022-06-12 [1] CRAN (R 4.4.0)
   benchmarkmeData            1.0.4      2020-04-23 [1] CRAN (R 4.4.0)
   Biobase                  * 2.64.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.1)
   BiocFileCache              2.12.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocGenerics             * 0.50.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocIO                     1.14.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocManager                1.30.23    2024-05-04 [1] CRAN (R 4.4.0)
   BiocNeighbors              1.22.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocParallel             * 1.38.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocSingular               1.20.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   BiocVersion                3.19.1     2024-04-22 [1] Bioconductor 3.19 (R 4.4.0)
   Biostrings                 2.72.1     2024-06-02 [1] Bioconductor 3.19 (R 4.4.0)
   bit                        4.0.5      2022-11-15 [1] CRAN (R 4.4.0)
   bit64                      4.0.5      2020-08-30 [1] CRAN (R 4.4.0)
   bitops                     1.0-7      2021-04-24 [1] CRAN (R 4.4.0)
   blob                       1.2.4      2023-03-17 [1] CRAN (R 4.4.0)
   boot                       1.3-30     2024-02-26 [1] CRAN (R 4.4.1)
   broom                      1.0.6      2024-05-17 [1] CRAN (R 4.4.0)
   bslib                      0.7.0      2024-03-29 [1] CRAN (R 4.4.0)
   cachem                     1.1.0      2024-05-16 [1] CRAN (R 4.4.0)
   caTools                    1.18.2     2021-03-28 [1] CRAN (R 4.4.0)
   class                      7.3-22     2023-05-03 [1] CRAN (R 4.4.1)
   classInt                   0.4-10     2023-09-05 [1] CRAN (R 4.4.0)
 P cli                        3.6.3      2024-06-21 [2] CRAN (R 4.4.0)
   codetools                  0.2-20     2024-03-31 [1] CRAN (R 4.4.1)
   colorout                 * 1.3-0.2    2024-05-01 [1] Github (jalvesaq/colorout@c6113a2)
   colorspace                 2.1-0      2023-01-23 [1] CRAN (R 4.4.0)
   config                     0.3.2      2023-08-30 [1] CRAN (R 4.4.0)
   corpcor                    1.6.10     2021-09-16 [1] CRAN (R 4.4.0)
   cowplot                    1.1.3      2024-01-22 [1] CRAN (R 4.4.0)
   crayon                     1.5.3      2024-06-20 [1] CRAN (R 4.4.0)
   curl                       5.2.1      2024-03-01 [1] CRAN (R 4.4.0)
   data.table               * 1.15.4     2024-03-30 [2] CRAN (R 4.4.0)
   DBI                        1.2.3      2024-06-02 [1] CRAN (R 4.4.0)
   dbplyr                     2.5.0      2024-03-19 [1] CRAN (R 4.4.0)
   DelayedArray               0.30.1     2024-05-07 [1] Bioconductor 3.19 (R 4.4.0)
   DelayedMatrixStats         1.26.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   deldir                     2.0-4      2024-02-28 [1] CRAN (R 4.4.0)
   digest                     0.6.36     2024-06-23 [1] CRAN (R 4.4.0)
   doParallel                 1.0.17     2022-02-07 [1] CRAN (R 4.4.1)
   dotCall64                  1.1-1      2023-11-28 [1] CRAN (R 4.4.0)
   dplyr                      1.1.4      2023-11-17 [1] CRAN (R 4.4.0)
   dqrng                      0.4.1      2024-05-28 [1] CRAN (R 4.4.0)
   dreamlet                 * 1.2.1      2024-06-09 [1] Bioconductor 3.19 (R 4.4.0)
   DropletUtils               1.24.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   DT                         0.33       2024-04-04 [1] CRAN (R 4.4.0)
   e1071                      1.7-14     2023-12-06 [1] CRAN (R 4.4.0)
   EBImage                    4.46.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   edgeR                    * 4.2.0      2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   EnrichmentBrowser          2.34.1     2024-05-06 [1] Bioconductor 3.19 (R 4.4.0)
   EnvStats                   2.8.1      2023-08-22 [1] CRAN (R 4.4.0)
   evaluate                   0.24.0     2024-06-10 [1] CRAN (R 4.4.0)
   ExperimentHub              2.12.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   fANCOVA                    0.6-1      2020-11-13 [1] CRAN (R 4.4.0)
   fansi                      1.0.6      2023-12-08 [1] CRAN (R 4.4.0)
   fastmap                    1.2.0      2024-05-15 [1] CRAN (R 4.4.0)
   fftwtools                  0.9-11     2021-03-01 [1] CRAN (R 4.4.0)
   fields                     16.2       2024-06-27 [1] CRAN (R 4.4.0)
   filelock                   1.0.3      2023-12-11 [1] CRAN (R 4.4.0)
   foreach                    1.5.2      2022-02-02 [1] CRAN (R 4.4.0)
   generics                   0.1.3      2022-07-05 [1] CRAN (R 4.4.0)
   GenomeInfoDb             * 1.40.1     2024-05-24 [1] Bioconductor 3.19 (R 4.4.0)
   GenomeInfoDbData           1.2.12     2024-05-01 [1] Bioconductor
   GenomicAlignments          1.40.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   GenomicRanges            * 1.56.1     2024-06-12 [1] Bioconductor 3.19 (R 4.4.1)
   ggbeeswarm                 0.7.2      2023-04-29 [1] CRAN (R 4.4.0)
   ggplot2                  * 3.5.1      2024-04-23 [1] CRAN (R 4.4.0)
   ggrepel                    0.9.5      2024-01-10 [1] CRAN (R 4.4.0)
   glue                       1.7.0      2024-01-09 [1] CRAN (R 4.4.0)
   golem                      0.4.1      2023-06-05 [1] CRAN (R 4.4.0)
   gplots                     3.1.3.1    2024-02-02 [1] CRAN (R 4.4.0)
   graph                      1.82.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   gridExtra                * 2.3        2017-09-09 [1] CRAN (R 4.4.0)
   GSEABase                   1.66.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   gtable                     0.3.5      2024-04-22 [1] CRAN (R 4.4.0)
   gtools                     3.9.5      2023-11-20 [1] CRAN (R 4.4.0)
   HDF5Array                  1.32.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   here                       1.0.1      2020-12-13 [1] CRAN (R 4.4.0)
   hms                        1.1.3      2023-03-21 [1] CRAN (R 4.4.0)
   htmltools                  0.5.8.1    2024-04-04 [1] CRAN (R 4.4.0)
   htmlwidgets                1.6.4      2023-12-06 [1] CRAN (R 4.4.0)
   httpuv                     1.6.15     2024-03-26 [1] CRAN (R 4.4.0)
   httr                       1.4.7      2023-08-15 [1] CRAN (R 4.4.0)
   invgamma                   1.1        2017-05-07 [1] CRAN (R 4.4.0)
   IRanges                  * 2.38.1     2024-07-03 [1] Bioconductor 3.19 (R 4.4.1)
   irlba                      2.3.5.1    2022-10-03 [1] CRAN (R 4.4.0)
   iterators                  1.0.14     2022-02-05 [1] CRAN (R 4.4.0)
   jpeg                       0.1-10     2022-11-29 [1] CRAN (R 4.4.0)
   jquerylib                  0.1.4      2021-04-26 [1] CRAN (R 4.4.0)
   jsonlite                   1.8.8      2023-12-04 [1] CRAN (R 4.4.0)
   KEGGgraph                  1.64.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   KEGGREST                   1.44.1     2024-06-19 [1] Bioconductor 3.19 (R 4.4.0)
   KernSmooth                 2.23-24    2024-05-17 [1] CRAN (R 4.4.1)
   knitr                      1.48       2024-07-07 [1] CRAN (R 4.4.1)
   later                      1.3.2      2023-12-06 [1] CRAN (R 4.4.0)
   lattice                    0.22-6     2024-03-20 [1] CRAN (R 4.4.1)
   lazyeval                   0.2.2      2019-03-15 [1] CRAN (R 4.4.0)
   lifecycle                  1.0.4      2023-11-07 [1] CRAN (R 4.4.0)
   limma                    * 3.60.3     2024-06-16 [1] Bioconductor 3.19 (R 4.4.0)
   lme4                       1.1-35.5   2024-07-03 [1] CRAN (R 4.4.1)
   lmerTest                   3.1-3      2020-10-23 [1] CRAN (R 4.4.0)
   locfit                     1.5-9.10   2024-06-24 [1] CRAN (R 4.4.0)
   magick                     2.8.3      2024-02-18 [1] CRAN (R 4.4.0)
   magrittr                   2.0.3      2022-03-30 [1] CRAN (R 4.4.0)
   maps                       3.4.2      2023-12-15 [1] CRAN (R 4.4.0)
   mashr                      0.2.79     2023-10-18 [1] CRAN (R 4.4.0)
   MASS                       7.3-61     2024-06-13 [1] CRAN (R 4.4.0)
   mathjaxr                   1.6-0      2022-02-28 [1] CRAN (R 4.4.0)
   Matrix                     1.7-0      2024-04-26 [1] CRAN (R 4.4.1)
   MatrixGenerics           * 1.16.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   matrixStats              * 1.3.0      2024-04-11 [1] CRAN (R 4.4.0)
   memoise                    2.0.1      2021-11-26 [1] CRAN (R 4.4.0)
   metadat                    1.2-0      2022-04-06 [1] CRAN (R 4.4.0)
   metafor                    4.6-0      2024-03-28 [1] CRAN (R 4.4.0)
   mime                       0.12       2021-09-28 [1] CRAN (R 4.4.0)
   minqa                      1.2.7      2024-05-20 [1] CRAN (R 4.4.0)
   mixsqp                     0.3-54     2023-12-20 [1] CRAN (R 4.4.0)
   msigdbr                    7.5.1      2022-03-30 [1] CRAN (R 4.4.0)
   munsell                    0.5.1      2024-04-01 [1] CRAN (R 4.4.0)
   mvtnorm                    1.2-5      2024-05-21 [1] CRAN (R 4.4.0)
   nlme                       3.1-165    2024-06-06 [1] CRAN (R 4.4.0)
   nloptr                     2.1.1      2024-06-25 [1] CRAN (R 4.4.0)
   numDeriv                   2016.8-1.1 2019-06-06 [1] CRAN (R 4.4.0)
   paletteer                  1.6.0      2024-01-21 [1] CRAN (R 4.4.0)
   parallelly               * 1.38.0     2024-07-27 [1] CRAN (R 4.4.0)
   pbkrtest                   0.5.3      2024-06-26 [1] CRAN (R 4.4.0)
   pillar                     1.9.0      2023-03-22 [1] CRAN (R 4.4.0)
   pkgconfig                  2.0.3      2019-09-22 [1] CRAN (R 4.4.0)
   plotly                     4.10.4     2024-01-13 [1] CRAN (R 4.4.0)
   plyr                       1.8.9      2023-10-02 [1] CRAN (R 4.4.0)
   png                        0.1-8      2022-11-29 [1] CRAN (R 4.4.0)
   prettyunits                1.2.0      2023-09-24 [1] CRAN (R 4.4.0)
   progress                   1.2.3      2023-12-06 [1] CRAN (R 4.4.0)
   promises                   1.3.0      2024-04-05 [1] CRAN (R 4.4.0)
   proxy                      0.4-27     2022-06-09 [1] CRAN (R 4.4.0)
   purrr                      1.0.2      2023-08-10 [1] CRAN (R 4.4.0)
   R.methodsS3                1.8.2      2022-06-13 [1] CRAN (R 4.4.0)
   R.oo                       1.26.0     2024-01-24 [1] CRAN (R 4.4.0)
   R.utils                    2.12.3     2023-11-18 [1] CRAN (R 4.4.0)
   R6                         2.5.1      2021-08-19 [1] CRAN (R 4.4.0)
   rappdirs                   0.3.3      2021-01-31 [1] CRAN (R 4.4.0)
   rbibutils                  2.2.16     2023-10-25 [1] CRAN (R 4.4.0)
   RColorBrewer               1.1-3      2022-04-03 [1] CRAN (R 4.4.0)
   Rcpp                       1.0.13     2024-07-17 [1] CRAN (R 4.4.0)
   RcppParallel               5.1.8      2024-07-06 [1] CRAN (R 4.4.0)
   RcppZiggurat               0.1.6      2020-10-20 [1] CRAN (R 4.4.0)
   RCurl                      1.98-1.14  2024-01-09 [1] CRAN (R 4.4.0)
   Rdpack                     2.6        2023-11-08 [1] CRAN (R 4.4.0)
   remaCor                    0.0.18     2024-02-08 [1] CRAN (R 4.4.0)
   rematch2                   2.1.2      2020-05-01 [1] CRAN (R 4.4.0)
   reshape2                   1.4.4      2020-04-09 [1] CRAN (R 4.4.0)
   restfulr                   0.0.15     2022-06-16 [1] CRAN (R 4.4.0)
   Rfast                      2.1.0      2023-11-09 [1] CRAN (R 4.4.0)
   Rgraphviz                  2.48.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   rhdf5                      2.48.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   rhdf5filters               1.16.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   Rhdf5lib                   1.26.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   RhpcBLASctl                0.23-42    2023-02-11 [1] CRAN (R 4.4.0)
   rjson                      0.2.21     2022-01-09 [1] CRAN (R 4.4.0)
 P rlang                    * 1.1.4      2024-06-04 [2] CRAN (R 4.4.1)
   rmarkdown                  2.27       2024-05-17 [1] CRAN (R 4.4.0)
   rmeta                      3.0        2018-03-20 [1] CRAN (R 4.4.0)
   rprojroot                  2.0.4      2023-11-05 [1] CRAN (R 4.4.0)
   Rsamtools                  2.20.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   RSQLite                    2.3.7      2024-05-27 [1] CRAN (R 4.4.0)
   rstudioapi                 0.16.0     2024-03-24 [1] CRAN (R 4.4.0)
   rsvd                       1.0.5      2021-04-16 [1] CRAN (R 4.4.0)
   rtracklayer                1.64.0     2024-05-06 [1] Bioconductor 3.19 (R 4.4.0)
   s2                         1.1.6      2023-12-19 [1] CRAN (R 4.4.0)
   S4Arrays                   1.4.1      2024-05-20 [1] Bioconductor 3.19 (R 4.4.0)
   S4Vectors                * 0.42.1     2024-07-03 [1] Bioconductor 3.19 (R 4.4.1)
   sass                       0.4.9      2024-03-15 [1] CRAN (R 4.4.0)
   ScaledMatrix               1.12.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   scales                     1.3.0      2023-11-28 [1] CRAN (R 4.4.0)
   scater                     1.32.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   scattermore                1.2        2023-06-12 [1] CRAN (R 4.4.0)
   scuttle                    1.14.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   sessioninfo                1.2.2      2021-12-06 [1] CRAN (R 4.4.0)
   sf                       * 1.0-16     2024-03-24 [1] CRAN (R 4.4.0)
   sfheaders                  0.4.4      2024-01-17 [1] CRAN (R 4.4.0)
   shiny                      1.8.1.1    2024-04-02 [1] CRAN (R 4.4.0)
   shinyWidgets               0.8.6      2024-04-24 [1] CRAN (R 4.4.0)
   SingleCellExperiment     * 1.26.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   sp                         2.1-4      2024-04-30 [1] CRAN (R 4.4.0)
   spam                       2.10-0     2023-10-23 [1] CRAN (R 4.4.0)
   SparseArray                1.4.8      2024-05-30 [1] Bioconductor 3.19 (R 4.4.0)
   sparseMatrixStats          1.16.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   SpatialExperiment        * 1.14.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   SpatialFeatureExperiment * 1.6.1      2024-05-15 [1] Bioconductor 3.19 (R 4.4.0)
   spatialLIBD              * 1.16.2     2024-05-28 [1] Bioconductor 3.19 (R 4.4.0)
   spData                     2.3.1      2024-05-31 [1] CRAN (R 4.4.0)
   spdep                      1.3-5      2024-06-10 [1] CRAN (R 4.4.0)
   SQUAREM                    2021.1     2021-01-13 [1] CRAN (R 4.4.0)
   statmod                    1.5.0      2023-01-06 [1] CRAN (R 4.4.0)
   stringi                    1.8.4      2024-05-06 [1] CRAN (R 4.4.0)
   stringr                    1.5.1      2023-11-14 [1] CRAN (R 4.4.0)
   SummarizedExperiment     * 1.34.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   terra                      1.7-78     2024-05-22 [1] CRAN (R 4.4.0)
   tibble                     3.2.1      2023-03-20 [1] CRAN (R 4.4.0)
   tidyr                      1.3.1      2024-01-24 [1] CRAN (R 4.4.0)
   tidyselect                 1.2.1      2024-03-11 [1] CRAN (R 4.4.0)
   tiff                       0.1-12     2023-11-28 [1] CRAN (R 4.4.0)
   truncnorm                  1.0-9      2023-03-20 [1] CRAN (R 4.4.0)
   UCSC.utils                 1.0.0      2024-05-06 [1] Bioconductor 3.19 (R 4.4.0)
   units                      0.8-5      2023-11-28 [1] CRAN (R 4.4.0)
   utf8                       1.2.4      2023-10-22 [1] CRAN (R 4.4.0)
   variancePartition        * 1.34.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   vctrs                      0.6.5      2023-12-01 [1] CRAN (R 4.4.0)
   vipor                      0.4.7      2023-12-18 [1] CRAN (R 4.4.0)
   viridis                    0.6.5      2024-01-29 [1] CRAN (R 4.4.0)
   viridisLite                0.4.2      2023-05-02 [1] CRAN (R 4.4.0)
   withr                      3.0.0      2024-01-16 [1] CRAN (R 4.4.0)
   wk                         0.9.2      2024-07-09 [1] CRAN (R 4.4.0)
   xfun                       0.45       2024-06-16 [1] CRAN (R 4.4.0)
   XML                        3.99-0.17  2024-06-25 [1] CRAN (R 4.3.3)
   xtable                     1.8-4      2019-04-21 [1] CRAN (R 4.4.0)
   XVector                    0.44.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   yaml                       2.3.9      2024-07-05 [1] CRAN (R 4.4.0)
   zeallot                    0.1.0      2018-01-28 [1] CRAN (R 4.4.0)
   zenith                     1.6.0      2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   zlibbioc                   1.50.0     2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)

 [1] /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library
 [2] /Users/bmulvey/Library/R/arm64/4.4/library

 P ── Loaded and on-disk path mismatch.

──────────────────────────────────────────────────────────────────────────────────
